+++ lib/widgets/auth_dialog.dart
@@ -1,10 +1,8 @@
-import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:histutor/utils/authentication.dart';
 
 import '../histutor.dart';
-import '../mainboard.dart';
 
 BuildContext context_;
 
@@ -99,7 +97,7 @@ class _GoogleButtonState extends State<GoogleButton> {
               Navigator.push(context,
                   MaterialPageRoute(
                    // fullscreenDialog: true,
-                    builder: (context) => mainboard(),
+                    builder: (context) => histutor(),
                   ),
               );
               // Navigator.of(context).pushReplacement(
